Language corr. (1)
l. 5: spelling of "indem"
l. 29: spelling of "Benutzer-Online-Liste" (I guess that's the technical correct term, isn't it? Please check on that)
l. 34: "Mitgliedern" (Dative-case)
l. 36: "Benutzerrängen" (plural prefix was added) as well as deleting the comma  in front of "sowie" ( (that's due to the last German spelling rules)
l. 87: including the comma due to focus
l. 98: "Signaturen", the /n" was missing