Language correction (1)
I've noticed several phrases where the definite article would be appropriate due to stylistic reasons. Accordingly, I've inserted "the" in these lines: 5, 30, 31, 32, 36.

Furthermore I've inserted "one's" as a reflexive determiner in the lines 20, 23 and 25 (to establish the connection between the subject and object).

Within the lines 27-29 you wrote "see" and I wanted to know whether "view" wouldn't be more appropriate within this context? I think it's more likely to assume users to view sth., not the real visual seeing of sth.If you agree, let me know and I'll change the verb. 

Then I've inserted the "with"-gerund-structure in line 36, in order to the reveal the logical consequence of the two sentence parts.